Add routes, sub pages and navigation
