+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "svelte-spa-router": "^4.0.1"
+  }
+}
