+++ unit-15/jn-svelte-app/src/NotFound.svelte
@@ -1,0 +1,10 @@
+<script>
+    import { link } from "svelte-spa-router";
+
+    let title = "Page Not Found"
+</script>
+
+<main>
+    <h1>Oops, {title}</h1>
+    <p>Back to <a href="/" use:link>Home Page</a></p>    
+</main>
